worker_service/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4
html2text==2020.1.16
fake-useragent==1.4.0
python-dateutil==2.8.2
certifi==2023.11.17
urllib3==2.1.0
redis==5.0.1
python-dotenv==1.0.0